Move code to Shared
Move all of the C# for the libraries into a Shared folder and link to each project from there.
