+++ src/Shared/ITestOutputHelperAccessor.cs
