+++ src/Shared/ITestOutputHelperExtensions.cs
