+++ src/Shared/StringSyntaxAttribute.cs
