+++ src/Shared/TestOutputHelperAccessor.cs
